% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactLocation.R
\name{calculate_cell_location}
\alias{calculate_cell_location}
\title{Title}
\usage{
calculate_cell_location(
  cell_spot_map,
  st_meta,
  sc_meta,
  sc_norm,
  st_norm,
  batch = TRUE,
  pca_method = "prcomp_irlba",
  parallel = FALSE,
  st_gene_list = NULL,
  spot_neigh_list,
  radius = 1/2
)
}
\arguments{
\item{radius}{}
}
\description{
Title
}
