% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_colocaliztion.R
\name{cell_type_colocalization}
\alias{cell_type_colocalization}
\title{Generate a standard dataframe to store the number of cell-type colocalizations}
\usage{
cell_type_colocalization(df, cell_type = "celltype_0916")
}
\arguments{
\item{df}{A dataframe of cells, containing position and annotation information.}

\item{cell_type}{The name of column that stores cell type annotations.}
}
\value{
A empty dataframe designed to store the number of cell-type colocalizations.
}
\description{
Generate a standard dataframe to store the number of cell-type colocalizations
}
