% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_colocaliztion.R
\name{celltype_colocalization_count}
\alias{celltype_colocalization_count}
\title{Count the number of cell-type colocalizations}
\usage{
celltype_colocalization_count(
  df,
  cut_distance = 2.5,
  x = "pred_loc_x",
  y = "pred_loc_y",
  cell_type = "celltype_0916"
)
}
\arguments{
\item{df}{A dataframe of cells, containing position and annotation information.}

\item{cut_distance}{The cutoff distance used to define cell colocaliztion.}

\item{x}{The name of the column storing the x-coordinate values.}

\item{y}{The name of the column storing the y-coordinate values.}

\item{cell_type}{The name of column storing cell type annotations.}
}
\value{
A dataframe storing the number of cell-type colocalizations.
}
\description{
Count the number of cell-type colocalizations
}
