% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{cm_names}
\alias{cm_names}
\title{Available named matplotlib palettes}
\usage{
cm_names()
}
\value{
a character vector of available palettes for \code{\link[=cm_pal]{cm_pal()}}
}
\description{
Available named matplotlib palettes
}
\examples{
image(volcano, col = cm_pal("terrain_r")(120))
}
