% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{cm_pal}
\alias{cm_pal}
\title{Create a function from a matplotlib palette}
\usage{
cm_pal(pal = "gist_earth")
}
\arguments{
\item{pal}{Name of matplotlib palette, see \code{\link[=cm_names]{cm_names()}} for available one}
}
\value{
a function like grey.colors
}
\description{
Create a function from a matplotlib palette
}
\examples{
image(volcano, col = cm_pal("terrain_r")(120))
}
