% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap-package.R
\docType{package}
\name{cmap-package}
\alias{cmap}
\alias{cmap-package}
\title{cmap: Provides colours from 'matplotlib' the Python package}
\description{
The names and functionality of palette generators from matplotlib.
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
