% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harmony_embedding.R
\name{cosine_normalize}
\alias{cosine_normalize}
\title{Cosine normalization}
\usage{
cosine_normalize(X, MARGIN = 1, do_safe = TRUE)
}
\arguments{
\item{X}{Matrix}

\item{MARGIN}{row (1) or column (2)}

\item{do_safe}{}
}
\description{
Cosine normalization
}
