% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictDomain.R
\name{data_to_transform}
\alias{data_to_transform}
\title{Title}
\usage{
data_to_transform(
  sc_norm,
  st_norm,
  spatial_genes,
  batch = TRUE,
  npc = 100,
  nclust = 100,
  pca_method = c("prcomp", "prcomp_irlba"),
  max.iter.harmony = 50,
  max.iter.cluster = 6,
  cosine_norm = FALSE
)
}
\arguments{
\item{cosine_norm}{}
}
\description{
Title
}
