% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harmony_embedding.R
\name{harmony_embedding}
\alias{harmony_embedding}
\title{Use harmony to remove the batch effect}
\usage{
harmony_embedding(
  st_norm,
  sc_norm,
  genes,
  batch = TRUE,
  npc = 100,
  nclust = 100,
  pca_method = c("prcomp", "prcomp_irlba"),
  max.iter.harmony = 50,
  max.iter.cluster = 6,
  cosine_norm = FALSE
)
}
\arguments{
\item{st_norm}{gene * spot}

\item{sc_norm}{gene * cell}

\item{nclust}{round(nrow(meta_data)/30)}

\item{cosine_norm}{}
}
\description{
Use harmony to remove the batch effect
}
