% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMRF_gene.R
\name{hmrf_spatial_gene}
\alias{hmrf_spatial_gene}
\title{Calculate the HMRF genes}
\usage{
hmrf_spatial_gene(
  spatial_obj,
  kmtest,
  k = 3,
  filter_hbb = TRUE,
  gene_sampling_from_top = 2500,
  use_spatial_genes = "binSpect",
  filter_method = "elbow",
  gene_samples = 500,
  spatial_network_name = "KNN_network"
)
}
\arguments{
\item{spatial_obj}{Hmrf_object}

\item{k}{The number of spatial domain; should be adjusted according to different datasets.}

\item{filter_hbb}{Filter the hbb related genes, TRUE or FALSE}

\item{gene_sampling_from_top}{Total spatial genes before sampling}

\item{use_spatial_genes}{Which of Giotto's spatial genes to use}

\item{filter_method}{Filter genes by top or by elbow method, prior to sampling}

\item{gene_samples}{The number of extracted genes}

\item{spatial_network_name}{Name of spatial network to use for HMRF}
}
\value{
Reference: https://rdrr.io/github/RubD/Giotto/src/R/python_hmrf.R
}
\description{
Calculate the HMRF genes
}
