% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMapToSpot.R
\name{map_spot_genes}
\alias{map_spot_genes}
\title{Calculate the spatial variable genes}
\usage{
map_spot_genes(sc_norm, sc_meta, st_norm, spatial_location)
}
\arguments{
\item{sc_norm}{A gene * cell matrix}

\item{sc_meta}{A dataframe of cells}

\item{spatial_location}{A dataframe of spots, column \code{HMRF_cluster} containing the spatial domain label.}
}
\description{
Calculate the spatial variable genes
}
