% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactLocation.R
\name{neighbor_spot}
\alias{neighbor_spot}
\title{Extract the neighboring spots for each spot}
\usage{
neighbor_spot(spot_neigh_list, top_sim_spot)
}
\arguments{
\item{top_sim_spot}{}
}
\description{
Extract the neighboring spots for each spot
}
