% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMRF_gene.R
\name{sampling_sp_genes}
\alias{sampling_sp_genes}
\title{Extract a specified number of HMRF genes}
\usage{
sampling_sp_genes(clust, sample_rate = 2, target = 500, seed = 10)
}
\arguments{
\item{target}{The target number of HMRF genes}

\item{seed}{}
}
\value{
Reference: https://rdrr.io/github/RubD/Giotto/src/R/python_hmrf.R
}
\description{
Extract a specified number of HMRF genes
}
