% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictDomain.R
\name{scale_data_by_column}
\alias{scale_data_by_column}
\title{Scale test data to the same scale with the traning data}
\usage{
scale_data_by_column(train_set, test_set)
}
\arguments{
\item{train_set}{spot*gene}

\item{test_set}{cell*gene}
}
\value{
A list including train and test matrics
}
\description{
Scale test data to the same scale with the traning data
}
