% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictDomain.R
\name{tune_parameter}
\alias{tune_parameter}
\title{Tune the parameters}
\usage{
tune_parameter(
  train_set,
  test_set,
  scale = TRUE,
  class.weight = TRUE,
  kernel = "radial",
  verbose = FALSE,
  cross_para = c(4, 6, 8, 10)
)
}
\arguments{
\item{train_set}{spot*(gene+label)}

\item{test_set}{cell*gene}

\item{scale}{whether scale data to \link{0,1}}

\item{class.weight}{whether considering data imbalance}

\item{cross_para}{}
}
\description{
Tune the parameters
}
