% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactLocation.R
\name{cor_spot_cell}
\alias{cor_spot_cell}
\title{Calculate the correlation between spots and cells}
\usage{
cor_spot_cell(sc_norm, st_norm, genes, batch = TRUE)
}
\arguments{
\item{batch}{}
}
\description{
Calculate the correlation between spots and cells
}
