% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMRF_gene.R
\name{hmrf_spatial_gene}
\alias{hmrf_spatial_gene}
\title{Title}
\usage{
hmrf_spatial_gene(
  spatial_obj,
  kmtest,
  k = 3,
  filter_hbb = TRUE,
  gene_sampling_from_top = 2500,
  use_spatial_genes = "binSpect",
  filter_method = "elbow",
  gene_samples = 500,
  spatial_network_name = "KNN_network"
)
}
\arguments{
\item{k}{: the number of spatial domain}

\item{spatial_network_name}{}

\item{spatial_obj:}{hmrf_object}

\item{filter_hbb:}{TRUE or FALSE}
}
\description{
Title
}
