% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMapToSpot.R
\name{map_cell_to_spot}
\alias{map_cell_to_spot}
\title{Version_3: SSIM_Map}
\usage{
map_cell_to_spot(
  sc_norm,
  sc_meta,
  st_norm,
  spatial_location,
  batch = TRUE,
  pca_method = "prcomp_irlba",
  pred_sc_svm,
  pred_st_svm,
  python_path,
  num_epochs = 2000L,
  para_distance = 1,
  para_density = 1
)
}
\arguments{
\item{pca_method}{}
}
\description{
Version_3: SSIM_Map
}
