% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMapToSpot.R
\name{map_spot_genes}
\alias{map_spot_genes}
\title{Calculate the spatial variable genes}
\usage{
map_spot_genes(sc_norm, sc_meta, st_norm, spatial_location)
}
\arguments{
\item{spatial_location}{}
}
\description{
Calculate the spatial variable genes
}
