% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_colocaliztion.R
\name{permute_cell_coordinates}
\alias{permute_cell_coordinates}
\title{Randomly shift the spatial localtion of each cell to a random position within 100 μm from its original location}
\usage{
permute_cell_coordinates(
  radius_permute = 25,
  df,
  x = "pred_loc_x",
  y = "pred_loc_y",
  cut_distance = 2.5,
  cell_type = "celltype_0916"
)
}
\arguments{
\item{radius_permute}{The radius of the random shift distance; should be adjusted according to different datasets.}

\item{df}{A dataframe of cells, containing position and annotation information.}

\item{x}{The name of the column storing the x-coordinate values.}

\item{y}{The name of the column storing the y-coordinate values.}

\item{cut_distance}{The cutoff distance used to define cell colocaliztion.}

\item{cell_type}{The name of column storing cell type annotations.}
}
\value{
A dataframe with the randomly shifted cell positions and corresponding colocalization counts.

Reference: zhang et al. (2023): https://github.com/ZhuangLab/whole_mouse_brain_MERFISH_atlas_scripts_2023/blob/main/scripts/cell_cell_contacts/randomize_and_count_cell_cell_contacts_15um.ipynb
In the lung cancer datasets, the coordinates are scaled such that a distance of 25 equals 100μm.
}
\description{
Randomly shift the spatial localtion of each cell to a random position within 100 μm from its original location
}
