% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictDomain.R
\name{PredictDomain}
\alias{PredictDomain}
\title{Predict the spatial domain}
\usage{
PredictDomain(
  train_set,
  test_set,
  scale = TRUE,
  class.weight = TRUE,
  cost = 1,
  gamma = 1/ncol(test_set),
  kernel = "radial",
  st_svm = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{train_set}{spot*(gene+label)}

\item{test_set}{cell*gene}

\item{scale}{whether scale data to \link{0,1}}

\item{class.weight}{whether considering data imbalance}

\item{st_svm}{}
}
\description{
Predict the spatial domain
}
