% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictDomain.R
\name{scale_data_by_column}
\alias{scale_data_by_column}
\title{Scale}
\usage{
scale_data_by_column(train_set, test_set)
}
\arguments{
\item{train_set}{spot*gene}

\item{test_set}{cell*gene}
}
\description{
Scale data to \link{0,1} as SVM input
}
