% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactLocation.R
\name{spatial_relation_all}
\alias{spatial_relation_all}
\title{Calculate the adjacent spots of central spots}
\usage{
spatial_relation_all(
  spatial_location,
  spatial_data_type = c("honeycomb", "square", "singlecell_level"),
  n_near_spot = NULL,
  dis_cut = NULL
)
}
\arguments{
\item{dis_cut}{}
}
\description{
Calculate the adjacent spots of central spots
}
